requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
jinja2==3.1.4
pymongo==4.8.0
pymilvus==2.4.4
sentence-transformers==3.0.1
requests==2.32.3
fuzzywuzzy==0.18.0
python-levenshtein==0.25.1
reportlab==4.2.2
plotly==5.22.0
pandas==2.2.2
python-dotenv==1.0.1
semanticscholar==0.9.0
habanero==1.2.0
openai==1.35.3
marshmallow==3.21.3
environs==9.5.0
biopython==1.81  # Para PubMed